pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = """
    Task Manager is a task management system similar to http://www.redmine.org/.
    It allows you to set tasks, assign performers and change their statuses.
    Registration and authentication are required to work with the system.
"""
authors = ["Aleksey94Dan <danilchenko.aleksey94@gmail.com>"]
packages = [
  { include = "task_manager" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.2"
python-dotenv = "^0.17.0"
django-bootstrap4 = "^3.0.1"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.6"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
django-filter = "^2.4.0"
rollbar = "^0.15.2"

[tool.poetry.dev-dependencies]
Faker = "^8.1.4"
wemake-python-styleguide = "^0.15.2"
ipython = "^7.24.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
